Revert "Patch for add prf tests"
