+++ .gitignore
@@ -14,7 +14,6 @@ __pycache__
 */cython_version.py
 htmlcov
 .coverage
-coverage.xml
 MANIFEST
 .ipynb_checkpoints
 benchmarks/_results
@@ -37,7 +36,6 @@ docs/_build
 dist
 build
 eggs
-.eggs
 parts
 bin
 var
@@ -54,5 +52,3 @@ distribute-*.tar.gz
 
 # Mac OSX
 .DS_Store
-
-sandbox/
