--- .travis.yml
@@ -1,16 +1,0 @@
-language: python
-python:
-  - 3.6
-before_install:
-  - sudo apt-get install -y python-numpy python-scipy
-  - pip install astropy
-  - pip install pyketools
-  - pip install autograd
-  - pip install codecov
-  - pip install pytest-cov
-# command to run tests
-script:
-  - pytest
-
-after_sucess:
-  - codecov
