+++ README.md
@@ -1,20 +1,3 @@
-# 🐙 Oktopus
-![travis-ci](https://travis-ci.org/mirca/octopus.svg?branch=master)
+# 🐙 Octupus
 
 A soft-bodied, eight-armed package for beautiful inference.
-
-To install ``oktopus``, run:
-
-```
-pip install oktopus
-```
-
-or
-
-```
-git clone https://github.com/mirca/octopus.git
-cd octopus
-pip install -e .
-```
-
-And you're all set!
