+++ octopus/__init__.py
@@ -1,7 +1,1 @@
-import os
-from .prior import *
-from .likelihood import *
-from .posterior import *
-
-# Where are the Kepler PRF files stored by default?
-DEFAULT_PRFDIR = os.path.expanduser('~/.pyke/kepler-prf-calibration-data/')
+from .core import *
