--- setup.cfg
@@ -1,5 +1,0 @@
-[aliases]
-test=pytest
-
-[tool:pytest]
-addopts = -x --doctest-modules --ignore=setup.py --cov=octopus
